+++ src/main/java/io/github/astrapi69/time/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * The base package for helper and utility classes for the Date/Time API
+ * 
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.time;